Separate view and logic for player
